requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1
reportlab==4.0.9
python-multipart==0.0.6
jinja2==3.1.2
watchdog==3.0.0
pydantic-settings==2.1.0